desviacion promedio

+++ src/main/java/com/aws/codestart/projecttemplates/controller/response/KpiClient.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class KpiClient {
-	private float oldAverage;
-	private float oldDesviation;
+	private double oldAverage;
+	private double oldDesviation;
 	
 	
 }
